requirements.txt
pandas
python-docx
Pillow